% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{abs.threshold}
\alias{abs.threshold}
\title{Subset data for thresholded absolute values}
\usage{
\method{abs}{threshold}(x, threshold = 0.5)
}
\arguments{
\item{x}{the data to be subset}

\item{threshold}{the value to be used for subsetting}
}
\value{
a smaller version of the original data
}
\description{
This function removes from the passed data \code{x} all entried that, in absolute value, are
less of the passed \code{threshold}.
}
