% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{call.MTGO}
\alias{call.MTGO}
\title{Calls for MTGO execution}
\usage{
call.MTGO(outfolder, verbose = TRUE)
}
\arguments{
\item{outfolder}{the folder where to read data and write results}

\item{verbose}{if TRUE errors and warnings are printed after the run completes}
}
\value{
nothing
}
\description{
This function invokes MTGO. You need to have created a MTGO parameter file (see \link{write.paramFile}),
an edge file (see \link{write.edges}) and a dictionary file (see \link{write.dictionary}). Files formats and
names are standard so it's better
to use the suggested functions to create them.
}
