% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetTest_redone.R
\name{clean.filename}
\alias{clean.filename}
\title{Clean a string so that it's a valid file name}
\usage{
clean.filename(old_filename)
}
\arguments{
\item{old_filename}{filename to be cleaned (without the full path!)}
}
\value{
cleaned filename
}
\description{
Removes all characters that are not: literals, numbers, underscore, dash, dot
}
\keyword{internal}
