% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{coexpr_propr}
\alias{coexpr_propr}
\title{Compute coexpression via propr::propr}
\usage{
coexpr_propr(geneExpression, verbose = FALSE, ...)
}
\arguments{
\item{geneExpression}{a Seurat dgCMatrix (from the @data field in a Seurat object)}

\item{verbose}{if TRUE the messages from \link{propr::propr} are not suppressed}

\item{...}{extra parameters are passed to \link{propr::propr}}
}
\value{
the @matrix field of a propr object
}
\description{
This function is a wrapper around \link{propr::propr} to compute coexpression starting from
a Seurat dgCMatrix (from the @data field in a Seurat object), containing the expression data.
This function returns only the @matrix field of the propr object.
}
