% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{cor_threshold}
\alias{cor_threshold}
\title{Compute correlation after absolute threshold filter}
\usage{
cor_threshold(x, y, threshold = NULL, use = "everything",
  method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a second numeric vector}

\item{threshold}{numeric value. If NA the func}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation of)
one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or
"pairwise.complete.obs".}

\item{method}{a character string indicating which correlation coefficient
(or covariance) is to be computed. One of "pearson" (default), "kendall", or
"spearman": can be abbreviated.}
}
\value{
a numeric array with the computed correlations
}
\description{
This function is similar to the standard \link{cor}, but values are first filtered
comparing them in absolute value to the passed \code{threshold}. For this reason
it is not possible to support matrices as input.
If \code{threshold} is NULL the function just falls back to standard correlation function.
Otherwise it removes all values which are, for at least one array and in absolute value,
less then passed threshold
}
