% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTGOsc.R
\name{dummy.function}
\alias{dummy.function}
\title{Dummy function for import management}
\usage{
dummy.function()
}
\value{
nothing
}
\description{
This function allows to have a single place where to place all the generic @import.
Usually only a couple are required, but for things used all over the places (like stats package)
there is no real single place where to put the import. It's cleaner in this way.
}
\keyword{internal}
