% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetTest_redone.R
\name{export.network.modules}
\alias{export.network.modules}
\title{Export functional modules network}
\usage{
export.network.modules(infolder, collapse.modules = TRUE)
}
\arguments{
\item{infolder}{the cluster folder}

\item{collapse.modules}{should all elements of a module be collapsed into a single node?}
}
\value{
a visNetwork network object
}
\description{
Functional modules for a single cluster, after MTGO run. You can either collapse the
nodes of the same functional module, or keep them separated.
}
