% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetTest_redone.R
\name{export.network.single.module}
\alias{export.network.single.module}
\title{Export single module network(s)}
\usage{
export.network.single.module(infolder, modules = NULL)
}
\arguments{
\item{infolder}{working dir after MTGO run}

\item{modules}{name of the module(s) to be exported, or NULL to export all of them}
}
\value{
a list of visNetwork networks, indexed by module name
}
\description{
Export single module network(s)
}
