% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{get.filenames}
\alias{get.filenames}
\title{List standard filenames}
\usage{
get.filenames(outfolder)
}
\arguments{
\item{outfolder}{the root folder to be used.}
}
\value{
a list with all standard names listed as fields
}
\description{
All MTGOsc files have standard naming. This folders returns a list of them, starting from
a root folder.
}
