% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{scale.free.threshold}
\alias{scale.free.threshold}
\title{Subset a coexpression network to maximise free scale fit}
\usage{
\method{scale}{free.threshold}(x, thresholds = seq(from = 0.1, to = 0.9, by =
  0.1), target.gamma = 2, verbose = TRUE)
}
\arguments{
\item{x}{a coexpression network, as returned by \link{write.coexpressionMatrix}}

\item{thresholds}{numeric array of thresholds to be considered}

\item{target.gamma}{for biological networks target gamma is somewhere between 2 and 3}

\item{verbose}{should extra informations be printed?}
}
\value{
a smaller version of the original matrix
}
\description{
This function removes edges from the passed network so that the resulting subnetwork
is maximally scale free. First, for each value in \code{thresholds} argument, a subnetwork
is created. Each subnetwork is obtained from the original \code{x} removing all edges
smaller, in absolute value, than the corresponding threshold.
Then each subnetwork is fitted a power law, and the one whose gamma is closest
to \code{target.gamma} is selected.
}
