% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{thinning_abs_threshold}
\alias{thinning_abs_threshold}
\title{Subset a coexpression network for thresholded absolute values}
\usage{
thinning_abs_threshold(x, threshold = 0.5)
}
\arguments{
\item{x}{a coexpression matrix, as returned by \link{write.coexpressionMatrix}}

\item{threshold}{the value to be used for subsetting}
}
\value{
a smaller version of the original matrix
}
\description{
This function removes from the passed coexpression network \code{x} all edges that,
in absolute value, are less of the passed \code{threshold}.
}
