% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{write.coexpressionMatrix}
\alias{write.coexpressionMatrix}
\title{Compute and save a coexpression matrix}
\usage{
write.coexpressionMatrix(geneExpression, outfolder, overwrite = FALSE,
  fun = cor, ...)
}
\arguments{
\item{geneExpression}{a Seurat dgCMatrix (from the @data field in a Seurat object), containing the expression data}

\item{outfolder}{the data folder where to save the results}

\item{overwrite}{boolean. Should an existing coexpression file be overwritten on disk?}

\item{fun}{the coexpression function, it will be fed two genes' expression arrays. Defaults to \link[stats]{cor}}

\item{...}{all extra parameters are passed to coexpression function}
}
\value{
the coexpression matrix, in long form
}
\description{
This function uses the passed coexpression function (default: Pearson's correlation \link[stats]{cor})
to compute a coexpression matrix between each pair of genes expression arrays. The matrix is
saved in the passed directory.
}
