% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{write.dictionary}
\alias{write.dictionary}
\title{Create and save a dictionary file}
\usage{
write.dictionary(genes, terms, outfolder, overwrite = FALSE)
}
\arguments{
\item{genes}{character array, list of genes}

\item{terms}{character array, a list of terms associated with the genes (same order, same size)}

\item{outfolder}{the data folder where to save the dictionary file}

\item{overwrite}{boolean. Should an existing dictionary file be overwritten on disk?}
}
\value{
a data frame with genes and terms as columns
}
\description{
This function takes to sets (genes and corresponding terms) and saves a dictionary
file on disk.
}
