% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{write.edges}
\alias{write.edges}
\title{Compute and save network edges}
\usage{
write.edges(coexpression, outfolder, keep.weights = TRUE,
  outfile.name = NULL, overwrite = TRUE, fun = abs.threshold, ...)
}
\arguments{
\item{coexpression}{a coexpression matrix, in long form, as the one returned by \link{write.coexpressionMatrix}}

\item{outfolder}{the data folder where to save the results}

\item{keep.weights}{boolean. If TRUE arc weights are kept. Defaults to FALSE.}

\item{outfile.name}{optional filename for coexpression file}

\item{overwrite}{boolean. Should an existing edge file be overwritten on disk?}

\item{fun}{edge filtering function, defaults to \link{abs.threshold}}

\item{...}{extra arguments are passed to edge filtering function}
}
\value{
f
}
\description{
This function uses the passed function ()
}
