% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{write.paramFile}
\alias{write.paramFile}
\title{Writes a MTGO parameter file on disk}
\usage{
write.paramFile(outfolder, overwrite = FALSE, MinSize = 3,
  MaxSize = 30)
}
\arguments{
\item{outfolder}{the data folder where to save the param file}

\item{overwrite}{boolean. Should an existing parameter file be overwritten on disk?}

\item{MinSize}{MTGO MinSize parameter}

\item{MaxSize}{MTGO MaxSize parameter}
}
\value{
nothing
}
\description{
This function creates and saves a MTGO parameter file on disk.
}
