% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{write.edges}
\alias{write.edges}
\title{Compute and save network edges}
\usage{
write.edges(coexpression, outfolder, keep.weights = TRUE, overwrite = TRUE,
  fun = abs.threshold, ...)
}
\arguments{
\item{coexpression}{a coexpression matrix, in long form, as the one returned by \link{write.coexpressionMatrix}}

\item{outfolder}{the data folder where to save the results}

\item{keep.weights}{boolean. If TRUE arc weights are kept. Defaults to FALSE.}

\item{overwrite}{boolean. Should an existing edge file be overwritten on disk?}

\item{fun}{edge filtering function, defaults to \link{abs.threshold}}

\item{...}{extra arguments are passed to edge filtering function}
}
\value{
a trimmed network
}
\description{
This function takes a network (in the form of a coexpression matrix) and trims some of the
edges using the passed function \code{fun}. The resulting network is both returned and
saved on disk.
}
